Continue splitting
