+++ src/components/ChannelRow/index.js
@@ -1,0 +1,1 @@
+export { ChannelRow } from "./ChannelRow";
