+++ src/components/CumulativeStatsCard/index.js
@@ -1,0 +1,1 @@
+export { CumulativeStatsCard } from "./CumulativeStatsCard";
